+++ test/OpenTelemetry.Tests/Metrics/MetricPointReclaimTests.cs
@@ -17,6 +17,8 @@
 using System.Collections.Concurrent;
 using System.Diagnostics.Metrics;
 using System.Reflection;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 using OpenTelemetry.Tests;
 using Xunit;
 
@@ -24,6 +26,80 @@ namespace OpenTelemetry.Metrics.Tests;
 
 public class MetricPointReclaimTests
 {
+    public const string ReclaimUnusedMetricPointsConfigKey = "OTEL_DOTNET_EXPERIMENTAL_METRICS_RECLAIM_UNUSED_METRIC_POINTS";
+
+    private readonly Dictionary<string, string> configurationData = new()
+    {
+        [ReclaimUnusedMetricPointsConfigKey] = "true",
+    };
+
+    private readonly IConfiguration configuration;
+
+    public MetricPointReclaimTests()
+    {
+        this.configuration = new ConfigurationBuilder()
+            .AddInMemoryCollection(this.configurationData)
+            .Build();
+    }
+
+    [Theory]
+    [InlineData("false", false)]
+    [InlineData("False", false)]
+    [InlineData("FALSE", false)]
+    [InlineData("true", true)]
+    [InlineData("True", true)]
+    [InlineData("TRUE", true)]
+    public void TestReclaimAttributeConfigWithEnvVar(string value, bool isReclaimAttributeKeySet)
+    {
+        // Clear the environment variable value first
+        Environment.SetEnvironmentVariable(ReclaimUnusedMetricPointsConfigKey, null);
+
+        // Set the environment variable to the value provided in the test input
+        Environment.SetEnvironmentVariable(ReclaimUnusedMetricPointsConfigKey, value);
+
+        var exportedItems = new List<Metric>();
+
+        var meter = new Meter(Utils.GetCurrentMethodName());
+
+        using var meterProvider = Sdk.CreateMeterProviderBuilder()
+            .AddMeter(meter.Name)
+            .AddInMemoryExporter(exportedItems)
+            .Build();
+
+        var meterProviderSdk = meterProvider as MeterProviderSdk;
+        Assert.Equal(isReclaimAttributeKeySet, meterProviderSdk.ShouldReclaimUnusedMetricPoints);
+    }
+
+    [Theory]
+    [InlineData("false", false)]
+    [InlineData("False", false)]
+    [InlineData("FALSE", false)]
+    [InlineData("true", true)]
+    [InlineData("True", true)]
+    [InlineData("TRUE", true)]
+    public void TestReclaimAttributeConfigWithOtherConfigProvider(string value, bool isReclaimAttributeKeySet)
+    {
+        var exportedItems = new List<Metric>();
+
+        var meter = new Meter(Utils.GetCurrentMethodName());
+
+        using var meterProvider = Sdk.CreateMeterProviderBuilder()
+            .ConfigureServices(services =>
+            {
+                var configuration = new ConfigurationBuilder()
+                .AddInMemoryCollection(new Dictionary<string, string> { [ReclaimUnusedMetricPointsConfigKey] = value })
+                .Build();
+
+                services.AddSingleton<IConfiguration>(configuration);
+            })
+            .AddMeter(meter.Name)
+            .AddInMemoryExporter(exportedItems)
+            .Build();
+
+        var meterProviderSdk = meterProvider as MeterProviderSdk;
+        Assert.Equal(isReclaimAttributeKeySet, meterProviderSdk.ShouldReclaimUnusedMetricPoints);
+    }
+
     [Theory]
     [InlineData(false)]
     [InlineData(true)]
@@ -42,6 +118,10 @@ public class MetricPointReclaimTests
         };
 
         using var meterProvider = Sdk.CreateMeterProviderBuilder()
+            .ConfigureServices(services =>
+            {
+                services.AddSingleton(this.configuration);
+            })
             .AddMeter(Utils.GetCurrentMethodName())
             .AddReader(metricReader)
             .Build();
@@ -131,6 +211,10 @@ public class MetricPointReclaimTests
         };
 
         using var meterProvider = Sdk.CreateMeterProviderBuilder()
+            .ConfigureServices(services =>
+            {
+                services.AddSingleton(this.configuration);
+            })
             .AddMeter(Utils.GetCurrentMethodName())
             .SetMaxMetricPointsPerMetricStream(10) // Set max MetricPoints limit to 5
             .AddReader(metricReader)
@@ -152,12 +236,12 @@ public class MetricPointReclaimTests
         {
             int numberOfMeasurements = 0;
             var random = new Random();
-            while (emitMetricWithNoDimension)
+            while (true)
             {
                 if (numberOfMeasurements < numberOfMeasurementsPerThread)
                 {
                     // Check for cases where a metric with no dimension is also emitted
-                    if (true)
+                    if (emitMetricWithNoDimension)
                     {
                         counter.Add(25);
                         Interlocked.Add(ref sum, 25);
@@ -196,12 +280,12 @@ public class MetricPointReclaimTests
         Assert.Equal(sum, exporter.Sum);
     }
 
-    private class ThreadArguments
+    private sealed class ThreadArguments
     {
         public int Counter;
     }
 
-    private class CustomExporter : BaseExporter<Metric>
+    private sealed class CustomExporter : BaseExporter<Metric>
     {
         public long Sum = 0;